pyproject.toml
[tool.poetry]
name = "zane"
version = "0.1.0"
description = ""
authors = ["Kevin H Wilson <khwilson@gmail.com>"]
readme = "README.md"
packages = [
    {from = "src", include = "ijf"}
]

[tool.poetry.scripts]
ijf = "ijf.cli:cli"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
